+++ tests/test_directory.py
@@ -4,11 +4,12 @@ from laktory.models import Directory
 
 root_dir = os.path.dirname(__file__)
 
+directory = Directory(path=".laktory/pipelines/")
+
 
 def test_directory():
-    d = Directory(path=".laktory/pipelines/")
-    assert d.path == ".laktory/pipelines/"
-    assert d.resource_name == "directory-laktory-pipelines"
+    assert directory.path == ".laktory/pipelines/"
+    assert directory.resource_name == "directory-laktory-pipelines"
 
 
 if __name__ == "__main__":