+++ tests/test_job.py
@@ -1,6 +1,8 @@
 import os
+
 from laktory.models import Job
 from laktory import pulumi_outputs
+
 
 root_dir = os.path.dirname(__file__)