+++ tests/test_sql_query.py
@@ -3,14 +3,15 @@ from laktory.models import SqlQuery
 
 root_dir = os.path.dirname(__file__)
 
+query = SqlQuery(
+    name="google-prices",
+    parent="/queries",
+    query="SELECT * FROM dev.finance.slv_stock_prices",
+    data_source_id="12345",
+)
+
 
 def test_sql_query():
-    query = SqlQuery(
-        name="google-prices",
-        parent="/queries",
-        query="SELECT * FROM dev.finance.slv_stock_prices",
-        data_source_id="12345",
-    )
     data = query.model_dump()
     print(query.resource_key)
     assert query.resource_key == "google-prices"